backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
pydantic[email]
python-jose[cryptography]
passlib[argon2]
argon2-cffi
celery[gevent]
redis
pytesseract
Pillow
PyPDF2
pdf2image
google-generativeai
websockets
python-multipart
pytest
httpx
pydantic-settings